+++ jave-example/src/main/java/ws/schild/jave/example/DefaultFFMPEGLocatorTest.java
@@ -30,8 +30,8 @@ public class DefaultFFMPEGLocatorTest {
     public static void main(String [] args)
     {        // TODO review the generated test code and remove the default call to fail.
         DefaultFFMPEGLocator locator= new  DefaultFFMPEGLocator();
-        //String exePath= locator.getFFMPEGExecutablePath();
-        System.out.println("Done");
+        String exePath= locator.getFFMPEGExecutablePath();
+        System.out.println("ffmpeg executable found in <"+exePath+">");
     }
     
 }