+++ Changelog.md
@@ -4,6 +4,9 @@
 - **2.5.1-SNAPSHOT** 
    - Prepare for next dev cycle
    - Renamed native folder where the executables are to nativebin too, to prevent Java 9+ problems
+   - Allow to create a MultimediaObject with an URL instead of a file as conversion source
+   - Enhanced logging when ffmpeg executable could not be found in resources
+   - Make getFFMPEGExecutablePath() public to see what executable is beeing used
 - **2.5.0** 
    - Prepare for next dev cycle
    - Renamed build artifacts with native in the name to nativebin to avoid module conflicts in Java 9+